Further janitorial work cleaning up my ignorance.
