--- humblebundle/__init__.py
@@ -1,6 +1,0 @@
-__author__ = "Joel Pedraza"
-__copyright__ = "Copyright 2014, Joel Pedraza"
-__license__ = "MIT"
-__all__ = ['HumbleApi']
-
-from humblebundle.client import HumbleApi
